Tue, Aug 26, 2025, 11:56 PM +02:00

+++ interpreter/interpreter.py
@@ -3,10 +3,10 @@ x = int(x)
 z = int(z)
 match y:
     case "-":
-        print(x-z)
+        print(float(x-z))
     case "+":
-        print(x+z)
+        print(float(x+z))
     case "*":
-        print(x*z)
+        print(float(x*z))
     case "/":
-        print(x/z)
+        print(float(x/z))
